+++ CHANGES.md
@@ -4,7 +4,7 @@
 ## 0.4 (unreleased)
 
 
-- Nothing changed yet.
+- Added Sentry logging for crashed processes.
 
 
 ## 0.3 (2025-09-18)